progress cleaning up exchange
